feat(svg): editable svg files with live preview + zoom/pan
- New svg file-type handler: CodeMirror editing in edit/split mode,
  live rendered preview in preview/split mode
- SvgPreview renders in-memory content as a blob URL so edits
  reflect before save; ImageViewer's disk-based path unchanged
- Extract ZoomPanCanvas from ImageViewer: shared zoom/pan/pinch +
  toolbar; both ImageViewer and SvgPreview delegate to it
- Topbar: add 'svg' to SHOW_VIEW_MODE_FILE_TYPES so the
  split/edit/preview toggle appears for svg tabs
- image handler: drop 'svg' from extensions (svg handler wins it)
- SvgPreview uses default 'image-viewer' className to trigger
  .prev-body:has(.image-viewer) full-bleed CSS
