+++ apps/desktop/src/components/file-types/image/index.ts
@@ -4,7 +4,7 @@ import { getFileTypeIcon } from '@/components/icons/FileIcon';
 
 const handler: FileTypeHandler = {
   id: 'image',
-  extensions: ['png', 'jpg', 'jpeg', 'gif', 'webp', 'svg', 'bmp', 'ico'],
+  extensions: ['png', 'jpg', 'jpeg', 'gif', 'webp', 'bmp', 'ico'],
   icon: getFileTypeIcon('image'),
   supportedViewModes: ['preview'],
   needsFileContent: false,
