+++ apps/desktop/src/components/file-types/svg/index.ts
@@ -1,0 +1,16 @@
+import type { FileTypeHandler } from '../types';
+import { SvgPreview } from './SvgPreview';
+import { getFileTypeIcon } from '@/components/icons/FileIcon';
+
+const handler: FileTypeHandler = {
+  id: 'svg',
+  extensions: ['svg'],
+  icon: getFileTypeIcon('image'),
+  supportedViewModes: ['edit', 'preview', 'split'],
+  defaultViewMode: 'edit',
+  needsFileContent: true,
+  useCodeMirror: true,
+  Preview: SvgPreview,
+};
+
+export default handler;
